+++ android/src/main/java/com/engagelab/privates/flutter_plugin_engagelab/UserReceiver.java
@@ -4,6 +4,7 @@ import android.content.Context;
 
 import com.engagelab.privates.common.component.MTCommonReceiver;
 import com.engagelab.privates.push.api.AliasMessage;
+import com.engagelab.privates.push.api.CmdMessage;
 import com.engagelab.privates.push.api.CustomMessage;
 import com.engagelab.privates.push.api.InAppMessage;
 import com.engagelab.privates.push.api.NotificationMessage;
@@ -99,6 +100,15 @@ public class UserReceiver extends MTCommonReceiver {
     }
 
     /**
+     * 通用命令结果回调，小米订阅消息结果也通过此回调返回。
+     */
+    @Override
+    public void onCommandResult(Context context, CmdMessage cmdMessage) {
+        FlutterPluginEngagelabPlugin.logD(TAG, "onCommandResult:" + cmdMessage.toString());
+        FlutterPluginEngagelabPlugin.onCommonReceiver("onCommandResult", MsgToJson.cmdMessageToJson(cmdMessage));
+    }
+
+    /**
      * 厂商token消息回调
      *
      * @param context              不为空
@@ -148,4 +158,3 @@ public class UserReceiver extends MTCommonReceiver {
         FlutterPluginEngagelabPlugin.onCommonReceiver("onVoipMessage", MsgToJson.voipMessageToJson(voipDataMessage));
     }
 }
-